requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
torch==2.1.0
torchaudio==2.1.0
whisperx
transformers>=4.21.0
faster-whisper>=0.9.0
pyannote.audio>=3.1.0
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0
Werkzeug==2.3.7
